Replace many Strings with Texts
Originaly motivation for String over Text in many places was that they
are likely at the program boundary and will just be printed anyway,
so converting to Text is an unnecessary waste. However, we are still
doing a fair amount of string manipulation (e.g. pretty printing) and it
is hard to predict how exactly they will be used. It seems safer to default
to Text whenever we are storing it in a datatype.