Remove runPythiaHandler
runPythiaHandler was created for testing, as we wanted to capture
pythia's output. After the refactoring to effectful typeclasses
(7b01d17e95ba3732c86deeb2e541cf0671176e07), however, runPythiaHandler's
utility was dubious as its functionality was essentially replicable by
using runPythia with a custom

    MonadTerminal m => putTextLn :: Text -> m ()

handler. We thus remove it to simplify the API.