Add root PythiaException
We add back the root exception type PythiaException. We used to
have this, though it was removed in 1d0dd94 as a simplification.

However, it is certainly useful now, to avoid callstacks for known
exceptions (i.e. we only want to print callstacks for an 'unknown'
exception). Having a root type makes this significantly easier.